Resolve douple authentication error
When a user cancels out of an authentication process, Salesforce
can't exit the session. If the user attempts to create a new
session, the process fails and user was getting a ghost org
added. This fix prevents the ghost org from getting added and
warns the user to find the original session or wait 5 minutes until
the process timesout.